Package name changes and restructureing
