+++ .gitignore
@@ -2,5 +2,6 @@
 .vscode
 
 node_modules
+dist
 *.map
 *.log
