--- index.js
@@ -1,6 +1,0 @@
-
-const Joi = require('joi');
-
-module.exports = {
-	Joi,
-};
