+++ packages/yaatt-cli/bin/yaatt.js
@@ -1,5 +1,34 @@
 #!/usr/bin/env node
 
-require('@babel/register');
+const path = require('path');
+const glob = require('glob');
+const { flatten, map, compose } = require('ramda');
+const yargs = require('yargs');
 
-require('../src/init.js').default();
+const { runTestSuite } = require('@yaatt/core/test-runner');
+const { mapFutureSync, throwError } = require('@yaatt/utils');
+const { logError, logNewLine } = require('@yaatt/utils/logger');
+
+// TODO: Use Joi to validate arguments
+// TODO: Have more config passed via arguments
+const validateArgs = suitePaths => {
+	if(!suitePaths.length) {
+		return throwError('Need to specify path to test suite');
+	}
+
+	return suitePaths;
+};
+
+const importTestCase = require;
+
+const initTests = compose(
+	mapFutureSync(runTestSuite),
+	map(importTestCase),
+	map(path.resolve),
+	flatten,
+	map(glob.sync),
+	validateArgs,
+);
+
+initTests(yargs.argv._ || [])
+	.fork(logError, logNewLine);
