+++ packages/yaatt-core/package.json
@@ -3,7 +3,23 @@
   "version": "0.0.0",
   "description": "Core helpers to programatically run api tests",
   "main": "index.js",
-  "repository": "__",
+  "repository": "https://github.com/phenax/yaatt",
   "author": "Akshay Nair<phenax5@gmail.com>",
-  "license": "MIT"
+  "license": "MIT",
+  "dependencies": {
+    "ramda": "^0.25.0",
+    "joi": "^13.4.0",
+    "chalk": "^2.4.1",
+    "fluture": "^8.0.2",
+    "axios": "^0.18.0"
+  },
+  "devDependencies": {
+    "jest": "^23.1.0"
+  },
+  "jest": {
+    "testEnvironment": "node",
+    "transform": {
+      "^.+\\.jsx?$": "babel-jest"
+    }
+  }
 }
