+++ packages/yaatt-core/tests/Response.test.js
@@ -1,6 +1,6 @@
 
-import Response from '../src/Response';
-import { Joi } from '../index.js';
+import Response from '../Response';
+import { Joi } from 'joi';
 
 describe('Response', () => {
