+++ packages/yaatt-utils/src/create-class.js
